skill: refactor to dispatcher pattern with update-pr subcommand
Refactors from `cosmic-lua -l skill.pr update` to cleaner
`cosmic-lua -l skill update-pr` invocation pattern.

Creates a central skill dispatcher in lib/skill/init.lua that
can handle multiple subcommands, starting with update-pr.

Changes:
- lib/skill/init.lua - dispatcher that checks arg for subcommands
- lib/skill/pr.lua - removed, using build.pr directly
- lib/skill/cook.mk - updated to build init.lua instead of pr.lua
- lib/cosmic/test_skill_pr.lua - updated tests for new invocation
- lib/build/cook.mk - make update-pr uses `cosmic -l skill update-pr`
