+++ lib/cosmic/test_skill_pr.lua
@@ -1,4 +1,4 @@
--- test skill.pr module integration with cosmic binary
+-- test skill module integration with cosmic binary
 
 local lu = require("luaunit")
 local unix = require("cosmo.unix")
@@ -7,25 +7,25 @@ local spawn = require("cosmic.spawn")
 
 local cosmic = path.join(os.getenv("TEST_BIN_DIR"), "bin", "cosmic")
 
-TestSkillPr = {}
+TestSkill = {}
 
-function TestSkillPr:test_skill_pr_loads()
-  local ok, out = spawn({cosmic, "-l", "skill.pr", "-e", "print('loaded')"}):read()
-  lu.assertTrue(ok, "cosmic -l skill.pr failed to load")
+function TestSkill:test_skill_loads()
+  local ok, out = spawn({cosmic, "-l", "skill", "-e", "print('loaded')"}):read()
+  lu.assertTrue(ok, "cosmic -l skill failed to load")
   lu.assertStrContains(out, "loaded")
 end
 
-function TestSkillPr:test_skill_pr_update_outside_actions()
+function TestSkill:test_skill_update_pr_outside_actions()
   -- when not in GitHub Actions, should show help
   local env = unix.environ()
   env[#env + 1] = "GITHUB_ACTIONS=false"
 
-  local ok, out = spawn({cosmic, "-l", "skill.pr", "update"}, { env = env }):read()
-  lu.assertTrue(ok, "skill.pr update should not error outside GitHub Actions")
+  local ok, out = spawn({cosmic, "-l", "skill", "update-pr"}, { env = env }):read()
+  lu.assertTrue(ok, "skill update-pr should not error outside GitHub Actions")
   lu.assertStrContains(out, "Updates PR title and description")
 end
 
-function TestSkillPr:test_skill_pr_update_requires_token()
+function TestSkill:test_skill_update_pr_requires_token()
   -- in GitHub Actions without token, should fail
   local env = unix.environ()
   env[#env + 1] = "GITHUB_ACTIONS=true"
@@ -38,7 +38,7 @@ function TestSkillPr:test_skill_pr_update_requires_token()
     end
   end
 
-  local ok, out, exit_code = spawn({cosmic, "-l", "skill.pr", "update"}, { env = env }):read()
-  lu.assertFalse(ok, "skill.pr update should fail without GITHUB_TOKEN")
+  local ok, out, exit_code = spawn({cosmic, "-l", "skill", "update-pr"}, { env = env }):read()
+  lu.assertFalse(ok, "skill update-pr should fail without GITHUB_TOKEN")
   lu.assertEquals(exit_code, 1)
 end
